Convert the Name or the AGS of a Bundesland

Convert the Name or the AGS of a Bundesland

code_bundesland(
  sourcevar,
  origin = "name",
  destination = "ags",
  factor = FALSE
)

Arguments

sourcevar

Vector which contains the codes or names to be converted.

origin

The following options are available:

ags: AGS (default).

name: Bundesland name.

destination

The following options are available:

ags: Bundesland AGS (default).

iso: The Bundesland two-character abbreviation.

name: Bundesland name.

name_eng: Bundesland name in English.

factor

If TRUE returns ordered factor.

Value

A character vector.

Details

This function converts a string of Bundesland names into the AGS, the standardized (English) name, or the Bundesland abbreviation.

If origin="AGS", the first two digits will be used to identify a Bundesland. It is therefore important that sourcevar is supplied as a character vector with a leading zeros if applicable.

See also

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
data(btw_sn)

btw_sn %>% 
 mutate(bl=code_bundesland(district, origin="ags", 
     destination="name"))
#>     district year voters  valid      bl
#> 1      14161 1994 220347 141753 Sachsen
#> 2      14161 1998 221872 180143 Sachsen
#> 3      14161 2002 211051 155164 Sachsen
#> 4      14161 2005 207889 155857 Sachsen
#> 5      14166 1994  55989  34744 Sachsen
#> 6      14166 1998  59322  47771 Sachsen
#> 7      14166 2002  58824  42699 Sachsen
#> 8      14166 2005  58265  42748 Sachsen
#> 9      14167 1994  85553  52834 Sachsen
#> 10     14167 1998  87376  69161 Sachsen
#> 11     14167 2002  85156  59262 Sachsen
#> 12     14167 2005  84047  60038 Sachsen
#> 13     14171 1994  72576  48608 Sachsen
#> 14     14171 1998  73035  60843 Sachsen
#> 15     14171 2002  71783  52347 Sachsen
#> 16     14171 2005  70633  53020 Sachsen
#> 17     14173 1994 121309  79515 Sachsen
#> 18     14173 1998 116613  94473 Sachsen
#> 19     14173 2002 115425  83610 Sachsen
#> 20     14173 2005 114099  84997 Sachsen
#> 21     14177 1994 119977  76969 Sachsen
#> 22     14177 1998 123494 101378 Sachsen
#> 23     14177 2002 122493  90379 Sachsen
#> 24     14177 2005 121188  92084 Sachsen
#> 25     14178 1994 174147 113721 Sachsen
#> 26     14178 1998 169367 133767 Sachsen
#> 27     14178 2002 166411 120089 Sachsen
#> 28     14178 2005 163465 120722 Sachsen
#> 29     14181 1994  76948  52359 Sachsen
#> 30     14181 1998  76930  63605 Sachsen
#> 31     14181 2002  76481  55572 Sachsen
#> 32     14181 2005  75738  57521 Sachsen
#> 33     14182 1994 114143  76061 Sachsen
#> 34     14182 1998 113413  93383 Sachsen
#> 35     14182 2002 112054  82177 Sachsen
#> 36     14182 2005 110519  83335 Sachsen
#> 37     14188 1994  79142  52956 Sachsen
#> 38     14188 1998  77128  63963 Sachsen
#> 39     14188 2002  76952  56990 Sachsen
#> 40     14188 2005  75924  58081 Sachsen
#> 41     14191 1994 119540  76387 Sachsen
#> 42     14191 1998 117575  96313 Sachsen
#> 43     14191 2002 114254  82671 Sachsen
#> 44     14191 2005 111815  83684 Sachsen
#> 45     14193 1994 114994  70747 Sachsen
#> 46     14193 1998 111621  91247 Sachsen
#> 47     14193 2002 110237  80164 Sachsen
#> 48     14193 2005 109173  81855 Sachsen
#> 49     14262 1994 370242 224264 Sachsen
#> 50     14262 1998 383778 312788 Sachsen
#> 51     14262 2002 388631 287346 Sachsen
#> 52     14262 2005 398043 292565 Sachsen
#> 53     14263 1994  53379  31901 Sachsen
#> 54     14263 1998  51433  39124 Sachsen
#> 55     14263 2002  48708  33871 Sachsen
#> 56     14263 2005  47700  33550 Sachsen
#> 57     14264 1994  47856  27973 Sachsen
#> 58     14264 1998  44323  33690 Sachsen
#> 59     14264 2002  38923  27305 Sachsen
#> 60     14264 2005  36718  26166 Sachsen
#> 61     14272 1994 124986  83613 Sachsen
#> 62     14272 1998 127253 103710 Sachsen
#> 63     14272 2002 126590  92517 Sachsen
#> 64     14272 2005 126442  94767 Sachsen
#> 65     14280 1994 126799  81415 Sachsen
#> 66     14280 1998 123036  98219 Sachsen
#> 67     14280 2002 124359  90762 Sachsen
#> 68     14280 2005 125339  95590 Sachsen
#> 69     14284 1994  83030  52002 Sachsen
#> 70     14284 1998  85085  67629 Sachsen
#> 71     14284 2002  83421  59587 Sachsen
#> 72     14284 2005  81276  59623 Sachsen
#> 73     14285 1994  97689  59055 Sachsen
#> 74     14285 1998  99355  79218 Sachsen
#> 75     14285 2002  97895  68841 Sachsen
#> 76     14285 2005  96632  71352 Sachsen
#> 77     14286 1994 129277  83102 Sachsen
#> 78     14286 1998 127539 102118 Sachsen
#> 79     14286 2002 123894  87339 Sachsen
#> 80     14286 2005 121062  88394 Sachsen
#> 81     14287 1994 123709  80242 Sachsen
#> 82     14287 1998 120866  98459 Sachsen
#> 83     14287 2002 119190  84980 Sachsen
#> 84     14287 2005 117949  88943 Sachsen
#> 85     14290 1994  88017  56435 Sachsen
#> 86     14290 1998  98380  80765 Sachsen
#> 87     14290 2002 101271  74085 Sachsen
#> 88     14290 2005 101820  78292 Sachsen
#> 89     14292 1994 120551  82475 Sachsen
#> 90     14292 1998 123033 100556 Sachsen
#> 91     14292 2002 125207  90572 Sachsen
#> 92     14292 2005 125206  95934 Sachsen
#> 93     14365 1994 381185 217398 Sachsen
#> 94     14365 1998 387546 297392 Sachsen
#> 95     14365 2002 390484 285107 Sachsen
#> 96     14365 2005 399808 294329 Sachsen
#> 97     14374 1994  73224  44592 Sachsen
#> 98     14374 1998 100136  79537 Sachsen
#> 99     14374 2002 102477  72678 Sachsen
#> 100    14374 2005 102452  74743 Sachsen
#> 101    14375 1994  64707  36819 Sachsen
#> 102    14375 1998  64314  50944 Sachsen
#> 103    14375 2002  62709  43071 Sachsen
#> 104    14375 2005  61737  45155 Sachsen
#> 105    14379 1994 179918 110333 Sachsen
#> 106    14379 1998 128809 103118 Sachsen
#> 107    14379 2002 124870  91381 Sachsen
#> 108    14379 2005 124696  93076 Sachsen
#> 109    14383 1994  92757  56621 Sachsen
#> 110    14383 1998 107882  86578 Sachsen
#> 111    14383 2002 110294  79047 Sachsen
#> 112    14383 2005 110595  82486 Sachsen
#> 113    14389 1994  80075  48905 Sachsen
#> 114    14389 1998  81944  64473 Sachsen
#> 115    14389 2002  81951  57487 Sachsen
#> 116    14389 2005  81127  59157 Sachsen
#> 117    14511 2009 205652 134497 Sachsen
#> 118    14511 2013 201249 134528 Sachsen
#> 119    14511 2017 195929 145871 Sachsen
#> 120    14521 2009 321118 204141 Sachsen
#> 121    14521 2013 302267 207281 Sachsen
#> 122    14521 2017 288076 214490 Sachsen
#> 123    14522 2009 284421 181786 Sachsen
#> 124    14522 2013 268349 183652 Sachsen
#> 125    14522 2017 256511 190229 Sachsen
#> 126    14523 2009 213732 135054 Sachsen
#> 127    14523 2013 201934 136695 Sachsen
#> 128    14523 2017 193081 140861 Sachsen
#> 129    14524 2009 297512 187146 Sachsen
#> 130    14524 2013 281789 189680 Sachsen
#> 131    14524 2017 269048 194153 Sachsen
#> 132    14612 2009 421639 281513 Sachsen
#> 133    14612 2013 430136 314728 Sachsen
#> 134    14612 2017 429448 336654 Sachsen
#> 135    14625 2009 278070 178924 Sachsen
#> 136    14625 2013 262663 179614 Sachsen
#> 137    14625 2017 252642 188470 Sachsen
#> 138    14626 2009 238958 147147 Sachsen
#> 139    14626 2013 223614 149961 Sachsen
#> 140    14626 2017 213021 153720 Sachsen
#> 141    14627 2009 216629 137821 Sachsen
#> 142    14627 2013 207261 143189 Sachsen
#> 143    14627 2017 201310 150913 Sachsen
#> 144    14628 2009 215674 138960 Sachsen
#> 145    14628 2013 207752 144975 Sachsen
#> 146    14628 2017 202580 154375 Sachsen
#> 147    14713 2009 415259 271106 Sachsen
#> 148    14713 2013 430321 288981 Sachsen
#> 149    14713 2017 448372 333714 Sachsen
#> 150    14729 2009 231084 146308 Sachsen
#> 151    14729 2013 220400 149205 Sachsen
#> 152    14729 2017 215616 159391 Sachsen
#> 153    14730 2009 178447 107956 Sachsen
#> 154    14730 2013 168695 110163 Sachsen
#> 155    14730 2017 163916 116563 Sachsen